requirements.txt
httpx==0.28.1
isort==6.0.1
flake8==7.2.0
fastapi==0.124.0
pydantic2==1.1.15
pydantic-settings==2.12.0
uvicorn==0.35.0
backoff==2.2.1